Adding batched_copy to All Communicators
This commit does following adjustments:
1. moving the batched_copy from pure_nccl_communicator to the
base class of communicators, which enables batched_copy in all the
communicators.
2. A batched_copy parameter is added to the __init__ of all the
communicators.
3. The warning when using batched_copy on creating communicators
other than pure_nccl_communicator is removed.
