+++ chainermn/communicators/naive_communicator.py
@@ -4,8 +4,10 @@ from chainermn.communicators import mpi_communicator_base
 
 class NaiveCommunicator(mpi_communicator_base.MpiCommunicatorBase):
 
-    def __init__(self, mpi_comm):
+    def __init__(self, mpi_comm,
+                 batched_copy=False):
         super(NaiveCommunicator, self).__init__(mpi_comm)
+        self.batched_copy = batched_copy
 
     def multi_node_mean_grad(self, model, zero_fill=False):
         params = _memory_utility.extract_params_set_grad(model, zero_fill)
